+++ .github/workflows/ci.yml
@@ -21,7 +21,7 @@ jobs:
       fail-fast: false
       matrix:
         include:
-          - { PYTHON_VERSION: 'python=3.9', POLARS_VERSION: 'polars=0.14' }
+          - { PYTHON_VERSION: 'python=3.9', POLARS_VERSION: 'polars=0.14.30' }
           - { PYTHON_VERSION: 'python=3.9', POLARS_VERSION: 'polars=0.15' }
           - { PYTHON_VERSION: 'python=3.9', POLARS_VERSION: 'polars=0.16' }
           - { PYTHON_VERSION: 'python=3.9', POLARS_VERSION: 'polars=0.17' }
@@ -35,6 +35,8 @@ jobs:
         run: |
           curl -fsSL https://raw.githubusercontent.com/prefix-dev/pixi/main/install/install.sh | bash
       - name: Install dependencies
+        # TODO: make prettier once there are feature flags
+        # https://github.com/prefix-dev/pixi/issues/239
         run: |
           pixi add ${{ matrix.PYTHON_VERSION }} ${{ matrix.POLARS_VERSION }}
           pixi install
