+++ pyproject.toml
@@ -20,7 +20,7 @@ classifiers = [
     "Programming Language :: Python :: 3.11",
 ]
 dependencies = [
-    "polars >=0.18,<0.19",
+    "polars >=0.14.24,<0.19",
 ]
 
 [project.urls]
