+++ tests/test_parse_body.py
@@ -1,196 +1,19 @@
-# ruff: noqa
-
 import inspect
-import polars as pl
+
+import polars
 import pytest
 from hypothesis import given
+from hypothesis.strategies import integers
+from packaging.version import Version
+from polars import __version__ as _pl_version
 from polars.testing import assert_frame_equal
 from polars.testing.parametric import column, dataframes
-from hypothesis.strategies import integers
 
 from polarify import polarify, transform_func_to_new_source
 
+from .functions import functions, xfail_functions
 
-def signum(x):
-    s = 0
-    if x > 0:
-        s = 1
-    elif x < 0:
-        s = -1
-    return s
-
-
-def signum_no_default(x):
-    if x > 0:
-        return 1
-    elif x < 0:
-        return -1
-    return 0
-
-
-def early_return(x):
-    if x > 0:
-        return 1
-    return 0
-
-
-def assign_both_branches(x):
-    if x > 0:
-        s = 1
-    else:
-        s = -1
-    return s
-
-
-def unary_expr(x):
-    s = -x
-    return s
-
-
-def call_target_identity(x):
-    return x
-
-
-def call_expr(x):
-    k = x * 2
-    s = call_target_identity(k + 3)
-    return s
-
-
-def if_expr(x):
-    s = 1 if x > 0 else -1
-    return s
-
-
-def if_expr2(x):
-    s = 1 + (x if x > 0 else -1)
-    return s
-
-
-def if_expr3(x):
-    s = 1 + ((3 if x < 10 else 5) if x > 0 else -1)
-    return s
-
-
-def compare_expr(x):
-    if (0 < x) & (x < 10):
-        s = 1
-    else:
-        s = 2
-    return s
-
-
-def chained_compare_expr(x):
-    if 0 < x < 10:
-        s = 1
-    else:
-        s = 2
-    return s
-
-
-def test_chained_compare_fail():
-    with pytest.raises(ValueError):
-        polarify(chained_compare_expr)
-
-
-def walrus_expr(x):
-    if (y := x + 1) > 0:
-        s = 1
-    else:
-        s = -1
-    return s * y
-
-
-def multiple_if_else(x):
-    if x > 0:
-        s = 1
-    elif x < 0:
-        s = -1
-    else:
-        s = 0
-    return s
-
-
-def nested_if_else(x):
-    if x > 0:
-        if x > 1:
-            s = 2
-        else:
-            s = 1
-    elif x < 0:
-        s = -1
-    else:
-        s = 0
-    return s
-
-
-def nested_if_else_expr(x):
-    if x > 0:
-        s = 2 if x > 1 else 1
-    elif x < 0:
-        s = -1
-    else:
-        s = 0
-    return s
-
-
-def assignments_inside_branch(x):
-    if x > 0:
-        s = 1
-        s = s + 1
-        s = x * s
-    elif x < 0:
-        s = -1
-        s = s - 1
-        s = x
-    else:
-        s = 0
-    return s
-
-
-def override_default(x):
-    s = 0
-    if x > 0:
-        s = 10
-    return x * s
-
-
-def no_if_else(x):
-    s = x * 10
-    k = x - 3
-    k = k * 2
-    return s * k
-
-
-def two_if_expr(x):
-    a = 1 if x > 0 else 5
-    b = 2 if x < 0 else 2
-    return a + b
-
-
-functions = [
-    signum,
-    early_return,
-    assign_both_branches,
-    unary_expr,
-    call_expr,
-    if_expr,
-    if_expr2,
-    if_expr3,
-    compare_expr,
-    multiple_if_else,
-    nested_if_else,
-    nested_if_else_expr,
-    assignments_inside_branch,
-    override_default,
-    no_if_else,
-    two_if_expr,
-]
-
-xfail_functions = [
-    walrus_expr,
-    signum_no_default,
-]
+pl_version = Version(_pl_version)
 
 
 @pytest.fixture(
@@ -208,19 +31,24 @@ def test_funcs(request):
     # build ast from transformed function as format as string
     transformed_func_unparsed = transform_func_to_new_source(original_func)
     print(
-        (
-            f"Original:\n{original_func_unparsed}\n"
-            f"Transformed:\n{transformed_func_unparsed}"
-        )
+        f"Original:\n{original_func_unparsed}\n"
+        f"Transformed:\n{transformed_func_unparsed}"
     )
     return transformed_func, original_func
 
 
+# chunking + apply is broken for polars < 0.18.1
+# https://github.com/pola-rs/polars/pull/9211
+# only relevant for our test setup, not for the library itself
 @given(
-    df=dataframes(column("x", dtype=pl.Int64, strategy=integers(-100, 100)), min_size=1)
+    df=dataframes(
+        column("x", dtype=polars.Int64, strategy=integers(-100, 100)),
+        min_size=1,
+        chunked=False if pl_version < Version("0.18.1") else None,
+    )
 )
-def test_transform_function(df: pl.DataFrame, test_funcs):
-    x = pl.col("x")
+def test_transform_function(df: polars.DataFrame, test_funcs):
+    x = polars.col("x")
     transformed_func, original_func = test_funcs
     assert_frame_equal(
         df.select(transformed_func(x).alias("apply")),
